Add downlaod helper and fix downlaod url
