+++ public/apps/index.html
@@ -70,48 +70,49 @@
     <![endif]-->
 
     <!-- Content -->
-
-    <!-- Dynamic content -->
-    <div id="root" content="none"></div>
-    <!-- /Dynamic content -->
-
     <div id="top" class="site-section">
       <div class="container">
 
-          <div class="downloads__content-container">
-            <div class="downloads__met-icon">
-              <img src="../images/metronome-apps-logo.png" />
-            </div>
-            <div class="downloads__title-container">
-              <h1 class="downloads__header-title">Download the App</h1>
-              <h2>Get started with the Metronome Wallet</h2>
-            </div>
-            <a class="btn">Download For Mac</a>
-            <a data-toggle="scroll__down" class="link__sm">See all download options</a>
-            <div class="downloads__app-screenshot">
-              <img class="--slideUp" src="../images/metronome-apps-demo.png" />
-            </div>
+        <div class="downloads__content-container">
+          <div class="downloads__met-icon">
+            <img src="../images/metronome-apps-logo.png" />
           </div>
-          <div class="envelope__footer">
-            <div class="envelope">
-              <div class="envelope__left"></div>
-              <div class="envelope__right"></div>
-            </div>
-            <h2>All Download Options</h2>
-            <div class="envelope__footer-downloads">
-              <ul>
-                <li><a class="btn">Download For Mac</a></li>
-                <li><a class="btn">Download For Windows</a></li>
-                <li><a class="btn">Download For Linux</a></li>
-              </ul>
-            </div>
-            <div data-toggle="scroll__here" class="envelope__footer-coming-soon">
-              <span class="span__strong">Coming soon</span> to the App Store and Google Play
-            </div>
-            </div>
+          <div class="downloads__title-container">
+            <h1 class="downloads__header-title">Download the App</h1>
+            <h2>Get started with the Metronome Wallet</h2>
+          </div>
+
+          <!-- Dynamic content -->
+          <div id="root" content="apps"></div>
+          <!-- /Dynamic content -->
+
+          <a data-toggle="scroll__down" class="link__sm">See all download options</a>
+          <div class="downloads__app-screenshot">
+            <img class="--slideUp" src="../images/metronome-apps-demo.png" />
+          </div>
+        </div>
+
+        <div class="envelope__footer">
+          <div class="envelope">
+            <div class="envelope__left"></div>
+            <div class="envelope__right"></div>
+          </div>
+
+          <h2>All Download Options</h2>
+          <div class="envelope__footer-downloads">
+            <ul>
+              <li><a class="btn">Download For Mac</a></li>
+              <li><a class="btn">Download For Windows</a></li>
+              <li><a class="btn">Download For Linux</a></li>
+            </ul>
+          </div>
+          <div data-toggle="scroll__here" class="envelope__footer-coming-soon">
+            <span class="span__strong">Coming soon</span> to the App Store and Google Play
+          </div>
+        </div>
+
       </div>
     </div>
-
     <!-- /Content -->
 
     <footer class="footer site-section" role="footer">
